groups/api/views.py
from rest_framework import generics
from groups.models import Group
from .serializers import GroupSerializer

class GroupListAPI(generics.ListCreateAPIView):
    queryset = Group.objects.all()
    serializer_class = GroupSerializer

class GroupDetailAPI(generics.RetrieveUpdateDestroyAPIView):
    queryset = Group.objects.all()
    serializer_class = GroupSerializer
