requirements.txt
cffi==1.13.2
Django==2.2.9
django-bootstrap3==12.0.3
django-bootstrap4==1.1.1
django-braces==1.13.0
djangorestframework==3.11.0
misaka==2.1.1
pycparser==2.19
Pygments==2.5.2
pytz==2019.3
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
